possiblement le dernier commit modification de la recuperation des exemples
