--- source/invoke/examples/check.py
@@ -1,11 +1,0 @@
-"""Module pour lancer les checks."""
-
-
-from invoke import run, task
-
-
-@task
-def checks(ctx):
-    """Check du code et de la doc."""
-    ctx.run("pycodestyle source")
-    ctx.run("pydocstyle source")
